revamp: ingredient's bulkactions in Bulkaction component
